requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
protobuf==3.20.1
python-decouple==3.6
python-dotenv==0.20.0
Werkzeug==2.2.2
zipp==3.8.1